refactor: rename from [lang] to [locale]
