+++ frontend/src/app/[locale]/globals.css
