+++ frontend/src/app/[locale]/translations.ts
